Req 3 ICar generica

+++ src/interfaces/IVehicle.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 
 const VehicleSchema = z.object({
   model: z.string().min(3),
-  year: z.number(),
+  year: z.number().int(),
   color: z.string().min(3),
   status: z.boolean().optional(),
   buyValue: z.number().int(),
